Add files via upload
This script is kind of wonky so you have to play with it some but you are able to get powershell to type anything including special characters, spaces, and return keys for you if you use what I have here. If you apply it to a hot key then it will send whatever message you want everytime although you might have to add a portion to make sure you have the correct window selected prior to typing otherwise that's up to the user.